==> notebooks/p1-2_preprocessing_yelpnyc.ipynb <==
# Preprocessing YelpNYC

# Set dataset name variable
ds_name = "YelpNYC"

## Import libraries

# Import libraries
import numpy as np # For numerical processing
import pandas as pd # For dataframe processing
from datetime import datetime # For date processing

## Process review metadata

### Read review metadata

# Read in review metadata 
df_meta = pd.read_csv("../datasets/raw/{0}/metadata".format(ds_name), sep="\t", header=None)
df_meta.columns = ["user", "product", "rating", "label", "date"]
df_meta.head()

# Print dataframe information
df_meta.info()

### Normalize data

# Convert date strings to datetime objects
df_meta.date = df_meta.date.apply(lambda value: str(datetime.strptime(value, "%Y-%m-%d").date()))  

# Convert rating to integer
df_meta.rating = df_meta.rating.astype(int)   

# Map label to category value
df_meta.label = df_meta.label.apply(lambda label: "fraud" if label == -1 else "organic")

df_meta.head()

## Process review content

### Read review content

# Read in review content
df_content = pd.read_csv("../datasets/raw/{0}/reviewContent".format(ds_name), sep="\t", header=None)
df_content.columns = ["user", "product", "date", "content"]
df_content.head()

df_content.info()

### Map content with (user ID, product Id and date), then check if there is duplicates

# Group by user, product, date
# Aggregate content into lists
content_mapping = list(df_content.groupby(["user", "product", "date"]).agg(list).to_dict().values())

# Print max number of reviews per user/product/date  
print(
    "Maximum number of review texts with the same user, product and date:", 
    max(map(lambda it: max(map(len, it.values())), content_mapping)),
)

# Flatten content mapping dictionary
content_mapping = {k: v[0] for it in content_mapping for k, v in it.items()}

## Compose final review data 

### Fill content with corresponding metadata to the final data

# Copy metadata dataframe
df_review = df_meta.copy()  

# Add content column filled with NaNs
df_review["content"] = np.nan
df_review = df_review[["content", *df_meta.columns]]

# Fill content by merging 
df_review.content = df_review.apply(
    lambda row: content_mapping[row[1], row[2], row[-1]] if (row[1], row[2], row[-1]) in content_mapping else np.nan,
    axis=1, 
)
df_review.head()

### Drop NaN data (if yes)

# Drop missing rows
df_review = df_review.dropna()

### Re-mapping User / Product ID

max_user_idx_length = len(str(df_review["user"].nunique()))
max_product_idx_length = len(str(df_review["product"].nunique()))

user_mapping = {
    user: "{}_user_{}".format(ds_name.lower(), str.zfill(str(idx), max_user_idx_length))
    for idx, user in enumerate(df_review["user"].unique())
}
product_mapping = {
    product: "{}_product_{}".format(ds_name.lower(), str.zfill(str(idx), max_product_idx_length))
    for idx, product in enumerate(df_review["product"].unique())
}

df_review["user"] = df_review["user"].apply(user_mapping.get)
df_review["product"] = df_review["product"].apply(product_mapping.get)
df_review.head()

## Export final data

# Print out DataFrame info to validate 
df_review.info()

# Export cleaned DataFrame to a CSV file
df_review.to_csv("../datasets/processed/{0}_reviews_preprocessed.csv".format(ds_name.lower()), index=False)
