==> notebooks/p2_sentiment_extraction.ipynb <==
## Import libraries

import gc
import ast
import json
import pickle
import ray
import numpy as np
import modin.pandas as mpd
import pandas as pd

from textblob import TextBlob
from tqdm import tqdm
from transformers import pipeline
from torch.utils.data import Dataset

ray.init(ignore_reinit_error=True)

## Define feature engineering process
- Read data 
- Calculate features
    - Word / Sentence tokenization
    - Sentiment (Polarity / Subjectivity)
    - Emotion classification
- Export data with calculation

def literal_eval(text):
    try:
        return ast.literal_eval(text)
    except:
        return np.nan
        

def process_calculation(ds_name):
    mdf = mpd.read_csv("../datasets/processed/{0}_reviews_preprocessed.csv".format(ds_name))
    mdf = mdf.merge(
        mdf.content.apply(
            lambda content: json.dumps({
                "polarity": np.round(TextBlob(content).sentiment.polarity, 3),   
                "subjectivity": np.round(TextBlob(content).sentiment.subjectivity, 3),
                "words": list(map(str, TextBlob(content).words)),
                "sentences": list(map(str, TextBlob(content).sentences)),
            },
        )), 
        left_index=True, 
        right_index=True,
    ).rename(columns={"content_x": "content_text", "content_y": "content_features"})
    mdf.to_csv("../datasets/processed/{0}_reviews_calculated.xlsx".format(ds_name), index=None, sep="\t")
    del mdf
    gc.collect()
    
    df = pd.read_csv("../datasets/processed/{0}_reviews_calculated.xlsx".format(ds_name), sep="\t")
    df["content_features"] = df["content_features"].apply(literal_eval)
    df = df.dropna()
    df = pd.concat(
        [df, pd.json_normalize(df["content_features"].tolist()).add_prefix('content_')], 
        axis=1,
    )
    df = df.drop(columns=["content_features"])
    df = df[["user", "product", "rating", "label", "date", *[col for col in df.columns if "content_" in col]]]
    df.to_excel("../datasets/processed/{0}_reviews_calculated.xlsx".format(ds_name), index=None)
    return

### Process emotion predictions

def create_emotion_predictions(ds_name):
  df = pd.read_excel("./drive/MyDrive/CMP1044/{}_reviews_calculated.xlsx".format(ds_name))


  class MyDataset(Dataset):
    inputs = df.content_text.values.tolist()
    def __len__(self):
      return len(self.inputs)

    def __getitem__(self, i):
      return str(self.inputs[i])


  pipe = pipeline("sentiment-analysis", model="michellejieli/emotion_text_classifier", device=0)
  dataset = MyDataset()

  model_outputs = []
  for out in tqdm(pipe(dataset, batch_size=64, truncation=True, padding=True, max_length=512), total=len(dataset)):
    model_outputs.append(out)
    if model_outputs and (len(model_outputs) % 1600 == 0 or len(model_outputs) >= len(dataset)):
      with open("./drive/MyDrive/CMP1044/{}_emotion_predictions.pkl".format(ds_name), "wb") as f:
        pickle.dump(model_outputs, f)


def get_emotion_predictions(ds_name):
    with open("../datasets/processed/{}_emotion_predictions.pkl".format(ds_name), "rb") as f:
        predictions = [it["label"] for it in pickle.load(f)]
    return predictions

## Conduct calculation and checkout

### YelpChi

ds_name = "yelpchi"
process_calculation(ds_name)

df = pd.read_excel("../datasets/processed/{0}_reviews_calculated.xlsx".format(ds_name))
df.head()

create_emotion_predictions("yelpchi")
df["content_emotion"] = get_emotion_predictions(ds_name)
df = df.dropna()
df.head()

df.info()
df.to_excel("../datasets/processed/{0}_reviews_calculated.xlsx".format(ds_name), index=None)

### YelpNYC

ds_name = "yelpnyc"
process_calculation(ds_name)

df = pd.read_excel("../datasets/processed/{0}_reviews_calculated.xlsx".format(ds_name))
df.head()

create_emotion_predictions("yelpnyc")
df["content_emotion"] = get_emotion_predictions(ds_name)
df = df.dropna()
df.head()

df.info()
df.to_excel("../datasets/processed/{0}_reviews_calculated.xlsx".format(ds_name), index=None)
