==> notebooks/p3_dumping_mongodb.ipynb <==
## Compose data

import os
import numpy as np
import pandas as pd

datasets = ["yelpchi", "yelpnyc"]

def read_ds(ds_name):
    result = pd.read_excel("../datasets/processed/{0}_reviews_calculated.xlsx".format(ds_name))
    result["dataset"] = ds_name
    result = result[[result.columns[-1], *result.columns[:-1]]]
    return result

df = pd.concat(
    [read_ds(ds_name) for ds_name in datasets], 
    ignore_index=True,
)

df.head()

df.info()

df = df.drop(columns=["content_sentences"])
df.content_words = df.content_words.apply(
    lambda content_words: list(map(lambda s: s.strip("'"), content_words.lstrip("[").rstrip("]").split(", ")))
)
df.info()

df.to_json("../datasets/processed/yelp_mongo_dumping.json", orient='records')

## Dump to MongoDB

data_path = os.path.abspath("../datasets/processed/yelp_mongo_dumping.json")

!docker stop mongo 
!docker volume remove mongodbdata && docker volume create mongodbdata
!docker run \
    -v mongodbdata:/data/db \
    -v {data_path}:/data/yelp_mongo_dumping.json \
    -p 27017:27017 \
    --rm -it -d --name mongo mongo

!docker exec -it mongo mongoimport \
    --uri "mongodb://localhost:27017/?retryWrites=true&w=majority" \
    --collection="yelp_reviews" \
    --db cmp1044 \
    --collection yelp_reviews \
    --file /data/yelp_mongo_dumping.json \
    --jsonArray

import urllib.parse
from pymongo import MongoClient
client = MongoClient("mongodb://localhost:27017/test")

collection = client.cmp1044.yelp_reviews
cursor = collection.find({})
df_dumped = pd.DataFrame(cursor)
df_dumped.head()
